seperate queryParameters into own module
