+++ client/src/main.js
@@ -4,17 +4,19 @@
 'use strict';
 global.$ = global.jQuery = require('jquery');
 const moment = require('moment-timezone');
-// global.Snap = require('snapsvg');
-const queryString = require('query-string');
 const SVGLoader = require('./js/svgloader');
 require('./js/ticker');
 
-const isTimezone = require('./js/isTimezone');
 const getCityFromTimezone = require('./js/getCityFromTimezone');
 
 const serviceUrl = '/data';
 const topStoriesUrl = serviceUrl + '/top-stories';
 const searchStoriesUrl = serviceUrl + '/search';
+
+const {
+	getQueryParams,
+	getQueryParam
+} = require('./js/queryParams.js');
 
 /*
 	Customisation
@@ -27,51 +29,6 @@ const searchStoriesUrl = serviceUrl + '/search';
 	&secondaryOffset=0
 	&secondaryMax=10
 */
-
-const {
-	getQueryParams,
-	getQueryParam
-} = (function () {
-	const parsed = queryString.parse(location.search);
-
-	const customTimezone = parsed.timezone;
-	const useCustomTimezone = isTimezone(customTimezone);
-
-	const primaryType = parsed.primaryType;
-	const primarySearch = parsed.primarySearch;
-	const primaryOffset = isNaN(parseInt(parsed.primaryOffset)) ? 0 : parseInt(parsed.primaryOffset);
-	const primaryMax = isNaN(parseInt(parsed.primaryMax)) ? 10 : parseInt(parsed.primaryMax);
-
-	const secondaryType = parsed.secondaryType;
-	const secondarySearch = parsed.secondarySearch;
-	const secondaryOffset = isNaN(parseInt(parsed.secondaryOffset)) ? 0 : parseInt(parsed.secondaryOffset);
-	const secondaryMax = isNaN(parseInt(parsed.secondaryMax)) ? 10 : parseInt(parsed.secondaryMax);
-
-	const defaultEdition = 'UK';
-	const edition = (parsed.edition) ? parsed.edition : defaultEdition;
-
-	const organisation = parsed.organisation;
-
-	const exportData = {
-		organisation,
-		edition,
-		customTimezone,
-		useCustomTimezone,
-		primaryMax,
-		primaryType,
-		primarySearch,
-		primaryOffset,
-		secondaryType,
-		secondarySearch,
-		secondaryOffset,
-		secondaryMax
-	};
-
-	return {
-		getQueryParams: () => exportData,
-		getQueryParam: i => exportData[i]
-	}
-}());
 
 function wait (ms) {
 	return new Promise(function (resolve) {
